Squashed commit of the following:
commit 6718585bf902c53a61df33a3301c883f6a8e1bb1
Author: Leabago <43474556+Leabago@users.noreply.github.com>
Date:   Thu Mar 13 15:35:04 2025 +0400

    disable delete button in archive state

commit b46e4ff60d711b55991a160155a2f19eeaf2e7c6
Author: Leabago <43474556+Leabago@users.noreply.github.com>
Date:   Thu Mar 13 15:34:39 2025 +0400

    disable delete button in archive state

commit 3db51dbd45e74f2e2d0d5d30f502b270e5595ad5
Author: Leabago <43474556+Leabago@users.noreply.github.com>
Date:   Thu Mar 13 14:27:55 2025 +0400

    rename PVC

commit 9af9e8ba8e8e95b8d1dc544037e71e118327ea37
Author: Leabago <43474556+Leabago@users.noreply.github.com>
Date:   Thu Mar 13 14:21:46 2025 +0400

    Disable uploading during archiving

commit 54551efc1a66fb4d5ae957094633bd6f40d54bbf
Author: Leabago <43474556+Leabago@users.noreply.github.com>
Date:   Thu Mar 13 12:24:59 2025 +0400

    add pvc storage

commit 28c50328658d5df6762babbb49ebfd338cc24cb7
Author: Leabago <43474556+Leabago@users.noreply.github.com>
Date:   Wed Mar 12 21:54:02 2025 +0400

    copy link with http

commit a92eaddf3ff01978d080a5c66704f680c5ed0e46
Author: Leabago <43474556+Leabago@users.noreply.github.com>
Date:   Wed Mar 12 21:21:01 2025 +0400

    set ddns var

commit ad3ecd99799aa4955c3ae8f8c110fa530f62ff0c
Author: Leabago <43474556+Leabago@users.noreply.github.com>
Date:   Wed Mar 12 19:05:43 2025 +0400

    redesign UI part

commit 86a070d155755d3fb55ebaeb1fe479924b6cd771
Author: Leabago <43474556+Leabago@users.noreply.github.com>
Date:   Fri Mar 7 20:18:57 2025 +0400

    update ui

commit 1b675e3c7327ceba620cd3a21dd4c56e8635f867
Author: Leabago <43474556+Leabago@users.noreply.github.com>
Date:   Thu Mar 6 22:42:32 2025 +0400

    align

commit a945ad459eff05ae6d16858a97341d0cddbb2536
Author: Leabago <43474556+Leabago@users.noreply.github.com>
Date:   Thu Mar 6 20:27:47 2025 +0400

    string MAX_FILE_SIZE

commit a37c7359ebc665677810d5ffff2e3455c7c5a8a1
Author: Leabago <43474556+Leabago@users.noreply.github.com>
Date:   Thu Mar 6 19:59:00 2025 +0400

    check maxFileSize

commit 50c3eba04a41110a538765e22f065425518bba5f
Author: Leabago <43474556+Leabago@users.noreply.github.com>
Date:   Thu Mar 6 15:40:13 2025 +0400

    duplicate file name

    duplicate file name

commit f2e35223a526e00275ed630aca8a30cd6b5221ca
Author: Leabago <43474556+Leabago@users.noreply.github.com>
Date:   Wed Mar 5 19:36:53 2025 +0400

    copyToClipboard

commit 0982fbccea970dd1602f80ea1f6be54ebe9f3598
Author: Leabago <43474556+Leabago@users.noreply.github.com>
Date:   Wed Mar 5 19:20:26 2025 +0400

    test copy

commit bd8b46eb233f63a68de7008fc99f602122d3c526
Author: Leabago <43474556+Leabago@users.noreply.github.com>
Date:   Wed Mar 5 19:03:44 2025 +0400

    test

+++ cmd/web/handlers.go
@@ -15,6 +15,7 @@ import (
 func ping(w http.ResponseWriter, r *http.Request) {
 	w.Write([]byte("Ok"))
 }
+
 func (app *application) createSnippetForm(w http.ResponseWriter, r *http.Request) {
 
 	// Generate a new session ID
@@ -23,6 +24,12 @@ func (app *application) createSnippetForm(w http.ResponseWriter, r *http.Request
 		app.logger.errorLog.Fatal(err)
 		return
 	}
+
+	mutex.Lock()
+	if _, exists := sessionMap[sessionID]; !exists {
+		sessionMap[sessionID] = &SessionState{IsArchiving: false}
+	}
+	mutex.Unlock()
 
 	// create unique user code assosiated with files
 	userCode := createUserCode()
@@ -46,7 +53,8 @@ func (app *application) createSnippetForm(w http.ResponseWriter, r *http.Request
 	http.SetCookie(w, &cookie)
 
 	app.render(w, r, "create.page.tmpl.html", &templateData{
-		Form: forms.New(nil),
+		Form:        forms.New(nil),
+		SessionCode: userCode,
 	})
 }
 
@@ -78,7 +86,7 @@ func (app *application) getSnippet(w http.ResponseWriter, r *http.Request) {
 		http.ServeFile(w, r, filePath)
 
 		// delete file if it is for one download
-		if file.OneDownload {
+		if file.OneTimeDownload {
 			app.redisClient.HDel(app.getRedisPath(availablePath, fileCode), fileInfoTitle)
 		}
 	} else {
@@ -89,17 +97,28 @@ func (app *application) getSnippet(w http.ResponseWriter, r *http.Request) {
 // homeGetFiles upload files to zip
 func (app *application) homeGetFiles(w http.ResponseWriter, r *http.Request) {
 
-	sessionID, userCode, err := app.getSessionValue(r)
+	sessionID, sessionCode, err := app.getSessionValue(r)
 	if err != nil {
 		app.serverError(w, err)
+		return
 	}
+
+	// check if the file is in Archiving status
+	mutex.Lock()
+	state := sessionMap[sessionID]
+	if state.IsArchiving {
+		mutex.Unlock()
+		app.serverErrorCode(w, fmt.Errorf("upload is prohibited while archive is in progress"), http.StatusForbidden)
+		return
+	}
+	mutex.Unlock()
 
 	var folderPathFull = filepath.Join(folderPath, sessionID)
 	app.logger.infoLog.Printf("create new folder %s", folderPathFull)
 
 	_, err = saveFilesToFolder(r, folderPathFull, app.maxFileSize)
 	if err != nil {
-		if errors.Is(err, fileTooLarge) {
+		if errors.Is(err, errFileTooLarge) {
 			app.fileTooLarge(w, err)
 		} else {
 			app.serverError(w, err)
@@ -107,7 +126,7 @@ func (app *application) homeGetFiles(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
-	w.Write([]byte(userCode))
+	w.Write([]byte(sessionCode))
 }
 
 func (app *application) redirectToArchive(w http.ResponseWriter, r *http.Request) {
@@ -136,16 +155,32 @@ func (app *application) archive(w http.ResponseWriter, r *http.Request) {
 	sessionID, fileCode, err := app.getSessionValue(r)
 	if err != nil {
 		app.serverError(w, err)
+		return
 	}
+
+	// check if the file is in Archiving status
+	mutex.Lock()
+	state := sessionMap[sessionID]
+	if state.IsArchiving {
+		mutex.Unlock()
+		app.serverErrorCode(w, fmt.Errorf("archive operation is already in progress"), http.StatusConflict)
+		return
+	}
+
+	// Set the archiving flag
+	state.IsArchiving = true
+	sessionMap[sessionID] = state
+	mutex.Unlock()
 
 	// create zip archive with files
 	fileNameList, err := createArhive(sessionID, fileCode)
 	if err != nil {
 		app.serverError(w, err)
+		return
 	}
 
 	// select file lifetime by selected radio value
-	lifeTime := selectLifeTime(r.FormValue("options"))
+	oneTimeDownload, lifeTime := selectLifeTime(r.FormValue("storageDuration"))
 
 	// collect file information
 	fullURL := getFullURL(r, fileCode)
@@ -153,29 +188,30 @@ func (app *application) archive(w http.ResponseWriter, r *http.Request) {
 	base64ImageData, err := createBase64ImageData(fullURL)
 	if err != nil {
 		app.serverError(w, err)
+		return
 	}
 
 	file := &models.File{
-		Name:         folderBegin + fileCode + zipName,
-		FileCode:     fileCode,
-		FileNameList: fileNameList,
-		OneDownload:  isOneDownload(r.FormValue("options")),
-		Exist:        true,
-		URL:          fullURL,
-		QRcodeBase64: base64ImageData,
+		Name:            folderBegin + fileCode + zipName,
+		FileCode:        fileCode,
+		FileNameList:    fileNameList,
+		OneTimeDownload: oneTimeDownload,
+		Exist:           true,
+		URL:             fullURL,
+		QRcodeBase64:    base64ImageData,
+		LifeTime:        lifeTime.String(),
 	}
 
 	fileJson, err := json.Marshal(file)
 	if err != nil {
 		app.serverError(w, err)
+		return
 	}
 
 	app.redisClient.HSet((app.getRedisPath(availablePath, fileCode)), fileInfoTitle, string(fileJson))
 	app.redisClient.Expire(app.getRedisPath(availablePath, fileCode), lifeTime).Result()
 
-	// redirect
-	redirectURL := filepath.Join("archive", fileCode)
-	http.Redirect(w, r, redirectURL, http.StatusSeeOther)
+	w.Write([]byte(fileCode))
 }
 
 // deleteOneFile delete only one file during session
@@ -184,10 +220,40 @@ func (app *application) deleteOneFile(w http.ResponseWriter, r *http.Request) {
 	sessionID, _, err := app.getSessionValue(r)
 	if err != nil {
 		app.serverError(w, err)
+		return
 	}
+
+	// check if the file is in Archiving status
+	mutex.Lock()
+	state := sessionMap[sessionID]
+	if state.IsArchiving {
+		mutex.Unlock()
+		app.serverErrorCode(w, fmt.Errorf("upload is prohibited while archive is in progress"), http.StatusForbidden)
+		return
+	}
+	mutex.Unlock()
+
 	fullPath := filepath.Join(folderPath, sessionID, fileName)
+
+	// Check if the file exists
+	if _, err := os.Stat(fileName); errors.Is(err, os.ErrNotExist) {
+		app.serverErrorCode(w, err, http.StatusNotFound)
+		return
+	}
+
 	err = os.Remove(fullPath)
 	if err != nil {
 		app.serverError(w, err)
+		return
 	}
+}
+
+func (app *application) getUserCode(w http.ResponseWriter, r *http.Request) {
+	_, fileCode, err := app.getSessionValue(r)
+	if err != nil {
+		app.serverError(w, err)
+		return
+	}
+
+	w.Write([]byte(fileCode))
 }
